Update link dir
- Update link dir
- Fix delete bug
- Add .srcjar

+++ scripts/delete_build_folders.py
@@ -25,6 +25,10 @@ def delete_build_folders(root_dir):
     for dirpath, dirnames, filenames in os.walk(root_dir):
         if 'build' in dirnames:
             build_dir = os.path.join(dirpath, 'build')
+            if "aosp-cmake" in build_dir:
+                print(f"Don't delete {build_dir}")
+                continue
+
             try:
                 shutil.rmtree(build_dir)
                 print(f"Deleted {build_dir}")
@@ -33,7 +37,8 @@ def delete_build_folders(root_dir):
 
 
 if __name__ == "__main__":
-    pwd = os.getcwd()
-    root_directory = os.path.dirname(pwd)
-
+    script_path = os.path.abspath(__file__)
+    script_directory = os.path.dirname(script_path)
+    root_directory = os.path.dirname(script_directory)
+    print(root_directory)
     delete_build_folders(root_directory)
